+++ pkg/controller/perconaservermongodb/finalizers.go
@@ -25,7 +25,7 @@ func (r *ReconcilePerconaServerMongoDB) checkFinalizers(ctx context.Context, cr 
 	log := logf.FromContext(ctx)
 
 	shouldReconcile = false
-	orderedFinalizers := naming.GetOrderedFinalizers(cr)
+	orderedFinalizers := GetOrderedFinalizers(cr)
 	var finalizers []string
 
 	for i, f := range orderedFinalizers {
@@ -273,4 +273,30 @@ func (r *ReconcilePerconaServerMongoDB) deleteSecrets(ctx context.Context, cr *a
 	}
 
 	return nil
+}
+
+func GetOrderedFinalizers(cr *api.PerconaServerMongoDB) []string {
+	order := []string{naming.FinalizerDeletePSMDBPodsInOrder, naming.FinalizerDeletePVC}
+
+	if cr.CompareVersion("1.17.0") < 0 {
+		order = []string{"delete-psmdb-pods-in-order", "delete-psmdb-pvc"}
+	}
+
+	finalizers := make([]string, len(cr.GetFinalizers()))
+	copy(finalizers, cr.GetFinalizers())
+	orderedFinalizers := make([]string, 0, len(finalizers))
+
+	for _, v := range order {
+		for i := 0; i < len(finalizers); {
+			if v == finalizers[i] {
+				orderedFinalizers = append(orderedFinalizers, v)
+				finalizers = append(finalizers[:i], finalizers[i+1:]...)
+				continue
+			}
+			i++
+		}
+	}
+
+	orderedFinalizers = append(orderedFinalizers, finalizers...)
+	return orderedFinalizers
 }
